% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsHydroParser.R
\name{UsHydroParser}
\alias{UsHydroParser}
\title{US hydrology MOPEX data parser}
\usage{
UsHydroParser(id)
}
\arguments{
\item{id}{main argument}
}
\description{
US hydrology MOPEX data parser
}
\examples{
UsHydroParser("01048000")
}
